fix broken build
Travis-ci reported that the tests had failed and therefore the build
was broken. This is caused by a commit that was intended to fail the
test and that means the tests are working. Hooray :nada:

Now it’s time to fix the build. This commit should pass all the tests
and be ready for deployment :rocket: