Custom tests using 'vows'
Created a custom test file and installed its dependencies.
Also expanded the `./test/travis-ci.sh` file to run this new test and
install the `vows` package globally.

The `vows` package is both a CLI package that runs the test file as
well as a module used to create the tests.